+++ package.json
@@ -20,7 +20,9 @@
     "react-router-dom": "6.2.2",
     "react-scripts": "5.0.0",
     "web-vitals": "2.1.4",
-    "xml2js": "0.4.23"
+    "xml2js": "0.4.23",
+    "xmldom": "0.6.0",
+    "xpath": "0.0.32"
   },
   "scripts": {
     "server": "node ./server",
