+++ server/services/RecipesService.js
@@ -1,13 +1,34 @@
 const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
 const parseStringPromise = require('xml2js').parseStringPromise;
 const _ = require('lodash');
+const NodeCache = require('node-cache');
+
+const recipesCache = new NodeCache({ stdTTL: 60 * 60 * 24 });
 
 const RECIPES_URL = 'https://www.dropbox.com/s/wclbvga9x50i0wv/recipes.xml?raw=1';
 const maxOptionLength = 15;		//// max length of filter option display value
 
 class RecipesService {
 
+  async getXml() {
+    const cache = recipesCache.get('xmlRecipes');
+    if (cache) {
+      return cache;
+    }
+    console.log('fetching xml');
+    const response = await fetch(RECIPES_URL);
+    const xml = await response.text();
+    recipesCache.set('xmlRecipes', xml);
+    return xml;
+  }
+
   async get() {
+    const cache = recipesCache.get('allRecipes');
+    if (cache) {
+      return cache;
+    }
+    console.log('building recipes')
+    // helpers
     const getKey = (object, key) => object[key] && object[key][0] ? object[key][0].trim() : '';
     const getMultiline = (object, key) => {
       if (object[key] && object[key][0]) {
@@ -18,8 +39,8 @@ class RecipesService {
         return [];
       }
     };
-    const response = await fetch(RECIPES_URL);
-    const xml = await response.text();
+    // let's go
+    const xml = await this.getXml();
     const { gourmetDoc : { recipe : xmlRecipes } } = await parseStringPromise(xml);
     let categories = [];
     let cuisines = [];
@@ -53,12 +74,14 @@ class RecipesService {
     categories = _.sortBy(_.uniqBy(categories, 'value'), o => o.display);
     cuisines = _.sortBy(_.uniqBy(cuisines, 'value'), o => o.display);
     ingredients = _.sortBy(_.uniqBy(ingredients, 'value'), o => o.display);
-    return {
+    const response = {
       categories,
       cuisines,
       ingredients,
       recipes
     };
+    recipesCache.set('allRecipes', response);
+    return response;
   }
 
   _getOption(option) {
