+++ src/index.js
@@ -3,8 +3,6 @@ import ReactDOM from 'react-dom';
 
 import App from './js/components/App';
 
-import './index.css';
-
 ReactDOM.render(
     <React.StrictMode>
       <App />
