+++ src/js/components/appview.css
@@ -1,4 +1,4 @@
-body {
+.app {
   margin: 0;
   padding: 0;
   background: black;
