+++ src/js/components/pages/recipes/Recipe.js
@@ -23,8 +23,8 @@ const recipeMetaMap = [
 
 const Links = ({ id }) => (
 	<span>
-		<a href={`/recipes/print/${id}`} target="_blank" rel="noreferrer">Print</a>&nbsp;|&nbsp;
-		<a href={`/recipes/export/${id}`} target="_blank" rel="noreferrer">Export</a>
+		<a href={`/print/recipes/${id}`} target="_blank" rel="noreferrer">Print</a>&nbsp;|&nbsp;
+		<a href={`/export/recipes/${id}`} target="_blank" rel="noreferrer">Export</a>
 		<br /><br />
 		<a href="#top">Return to top</a>
 	</span>
@@ -34,7 +34,7 @@ const Source = ({ recipe }) => {
   if (recipe.link && recipe.source) {
     return <div><dt>Source</dt><dd><a href={recipe.link} target="_blank" rel="noreferrer">{recipe.source}</a></dd></div>
   } else if (recipe.link) {
-    const display = recipe.link.replace(/\w+:\/\/([^\/]+).*/, '$1');
+    const display = recipe.link.replace(/\w+:\/\/([^/]+).*/, '$1');
     return <div><dt>Source</dt><dd><a href={recipe.link} target="_blank" rel="noreferrer">{display}</a></dd></div>
   } else if (recipe.source) {
     return <div><dt>Source</dt><dd>{recipe.source}</dd></div>
