+++ src/setupProxy.js
@@ -1,5 +1,5 @@
 const {createProxyMiddleware} = require('http-proxy-middleware');
 
 module.exports = function(app) {
-  app.use('/api', createProxyMiddleware({ target: process.env.REACT_APP_SERVER_URL, pathRewrite: {'^/api' : ''} }));
+  app.use(['/api', '/export'], createProxyMiddleware({ target: process.env.REACT_APP_SERVER_URL  }));
 };
